+++ Changelog.md
@@ -3,6 +3,7 @@
 ## 1.8.0
 
 - Provide folder selection when using `Create Camel Quarkus/SpringBoot Project` command
+- Avoid being notified for package.json opening and close as no support is provided for it and every 2 minutes there is a notification for these files
 
 ## 1.7.0